pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "data-eng-genai-demo"
version = "0.1.0"
description = "Pipeline de dados com Pandas — Top 10 Clientes de e-commerce"
requires-python = ">=3.9"
dependencies = [
    "pandas>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "flake8>=6.0",
    "build>=0.10",
]

[tool.black]
line-length = 79
target-version = ["py39"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
